rename: JwtToken이라 되있는 부분을 Jwt로 변경
