+++ src/main/java/com/fastcampus/befinal/common/response/code/JwtCode.java
@@ -3,12 +3,12 @@ package com.fastcampus.befinal.common.response.code;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public enum JwtTokenCode implements Code {
+public enum JwtCode implements Code {
     //error
-    NOT_VALID_JWT_TOKEN(1050),
-    EXPIRED_JWT_TOKEN(1051),
-    UNSUPPORTED_JWT_TOKEN(1052),
-    ILLEGAL_JWT_TOKEN(1053);
+    NOT_VALID_JWT(1050),
+    EXPIRED_JWT(1051),
+    UNSUPPORTED_JWT(1052),
+    ILLEGAL_JWT(1053);
 
     private final Integer code;
